support additional calllables with extra_callables
